整理代码风格、增加tiobe爬虫、template功能结构

+++ hellogithub/hellogithub/static/js/manage_category.js
@@ -1,17 +1,17 @@
 /**
  * Created by XueWeiHan on 17/7/14 下午11:54.
  */
+var csrftoken = $('meta[name=csrf-token]').attr('content');
+$.ajaxSetup({
+    beforeSend: function(xhr, settings) {
+        if (!/^(GET|HEAD|OPTIONS|TRACE)$/i.test(settings.type)) {
+            xhr.setRequestHeader("X-CSRFToken", csrftoken)
+        }
+    }
+});
+
 // {# 新增分类 #}
 $(document).ready(function() {
-    var csrftoken = $('meta[name=csrf-token]').attr('content');
-
-    $.ajaxSetup({
-        beforeSend: function(xhr, settings) {
-            if (!/^(GET|HEAD|OPTIONS|TRACE)$/i.test(settings.type)) {
-                xhr.setRequestHeader("X-CSRFToken", csrftoken)
-            }
-        }
-    });
     $("#create-category-submit").click(function() {
         $.ajax({
             url: '/manage/category/',
@@ -20,12 +20,10 @@ $(document).ready(function() {
                 category_name: $("#create-category-name").val()
             },
             success: function (result) {
-                if (result.code == 200){
-                    $("#result").html(result.message);
-                }
-                else {
-                    alert(result.message);
-                }
+                $("#result").html(result.message);
+            },
+            error: function (jqXHR) {
+                alert(jqXHR.responseJSON.message);
             }
         });
     });
@@ -42,11 +40,11 @@ $(document).on("click", "#edit-category-button", function() {
                 "<fieldset class=\"pure-group\">"+
                     "<div class=\"pure-control-group\">"+
                         "<label for=\"category-name\">Category Name：</label>"+
-                        "<input id=\"category-name\" value=\""+result.message.name+"\" type=\"text\" class=\"pure-input-1-2\">"+
+                        "<input id=\"category-name\" value=\""+result.payload.name+"\" type=\"text\" class=\"pure-input-1-2\">"+
                     "</div>"+
                 "</fieldset>"+
 
-                "<button id=\"edit-category-submit\" value=\""+result.message.id+"\" type=\"button\" class=\"pure-button pure-input-1 pure-button-primary\">Submit</button>"+
+                "<button id=\"edit-category-submit\" value=\""+result.payload.id+"\" type=\"button\" class=\"pure-button pure-input-1 pure-button-primary\">Submit</button>"+
             "</form>");
         }
     });
@@ -54,15 +52,6 @@ $(document).on("click", "#edit-category-button", function() {
 
 // {# 编辑 category #}
 $(document).on("click", "#edit-category-submit", function() {
-    var csrftoken = $('meta[name=csrf-token]').attr('content');
-
-    $.ajaxSetup({
-        beforeSend: function(xhr, settings) {
-            if (!/^(GET|HEAD|OPTIONS|TRACE)$/i.test(settings.type)) {
-                xhr.setRequestHeader("X-CSRFToken", csrftoken)
-            }
-        }
-    });
     $.ajax({
         url: "/manage/category/",
         type: "PUT",
@@ -71,44 +60,33 @@ $(document).on("click", "#edit-category-submit", function() {
             category_name: $("#category-name").val()
         },
         success: function(result) {
-            if (result.code == 200){
-                $("#result").html(result.message);
-            }
-            else {
-                alert(result.message);
-            }
+            $("#result").html(result.message);
+        },
+        error: function (jqXHR) {
+           alert(jqXHR.responseJSON.message);
         }
     });
 });
 
 // {# 删除 category #}
 $(document).on("click", "#delete-category-submit", function() {
-    var csrftoken = $('meta[name=csrf-token]').attr('content');
-
-    $.ajaxSetup({
-        beforeSend: function(xhr, settings) {
-            if (!/^(GET|HEAD|OPTIONS|TRACE)$/i.test(settings.type)) {
-                xhr.setRequestHeader("X-CSRFToken", csrftoken)
-            }
-        }
-    });
     var category_name = $(this).parent().siblings("#category-name").text();
     var category_id = $(this).parent().siblings("#category-id").text();
-    $.ajax({
-        url: "/manage/category/",
-        type: "DELETE",
-        data: {
-            category_name: category_name,
-            category_id: category_id
-        },
-        success: function (result) {
-            if (result.code == 200){
+    var user_choice = window.confirm("确定删除分类："+ category_name+"?");
+    if(user_choice){
+        $.ajax({
+            url: "/manage/category/",
+            type: "DELETE",
+            data: {
+                category_name: category_name,
+                category_id: category_id
+            },
+            success: function (result) {
                 $("#result").html(result.message);
-            }
-            else {
-                alert(result.message);
-            }
-        }
-    });
+            },
+            error: function (jqXHR) {
+               alert(jqXHR.responseJSON.message);
+            }});
+    }
 });
